% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageDensity.R
\name{ageDensity}
\alias{ageDensity}
\title{Probability Distribution of Species Age}
\usage{
ageDensity(t, lambda, mu, rho = 1)
}
\arguments{
\item{t}{branching time}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{rho}{sampling fraction (Default: 1.0)}
}
\description{
This function provides the probability for a species age
given speciation and extinction rate, and the sampling fraction
}
\details{
Some additional details
}
\examples{
ageDensity(lambda = 0.2, mu = 0.1, t = 3.0, rho = 0.7)

}
\seealso{
\code{\link{simSpAge}} for stochastic age simulation
}
\author{
Daniele Silvestro
}
