% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTipAges.R
\name{calculateTipAges}
\alias{calculateTipAges}
\title{Get Phylogenetic Ages for the Tips of a Phylogeny}
\usage{
calculateTipAges(tree)
}
\description{
Get Phylogenetic Ages for the Tips of a Phylogeny
}
