% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateBD.R
\name{estimateBD}
\alias{estimateBD}
\title{Esitmate Speciation and Extinction rate from phylogeny}
\usage{
estimateBD(phy, epsilon = 0, rho = 1, ml_optim = "subplex")
}
\arguments{
\item{phy}{An ultrametric bifurcating phylogenetic tree, in ape "phylo" format.}

\item{epsilon}{Extinction fraction}

\item{rho}{Sampling fraction}

\item{ml_optim}{Method to use for optimisation (Default: "subplex"). May be one of "optim", "subplex", "nlminb", "nlm" (partial unambigious string is allowed).}
}
\value{
A named vector of two parameters, lambda and mu.
}
\description{
This function estimates speciation and extinction rate from a phylogeny under a given extinction and sampling fraction
}
\examples{
estimateBD(phy = , epsilon = 0.5)
}
\author{
Torsten Hauffe
}
