% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanAge.R
\name{meanAge}
\alias{meanAge}
\title{Expected Mean Species Age}
\usage{
meanAge(v, lambda, mu, rho = 1)
}
\arguments{
\item{v}{branching time}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{rho}{sampling fraction (Default: 1.0)}
}
\description{
This function calculates the expected mean species age.
}
\examples{
meanAge(v = 3.0, lambda = 0.2, mu = 0.1, rho = 0.7)
}
\author{
Daniele Silvestro
}
