% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianAge.R
\name{medianAge}
\alias{medianAge}
\title{Expected Median Species Age}
\usage{
medianAge(lambda, mu, v, rho)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{v}{branching time}

\item{rho}{sampling fraction (Default: 1.0)}
}
\description{
This function calculates the expected median species age.
}
\examples{
medianAge(v = 5.0, lambda = 0.3, mu = 0.1, rho = 0.7)
}
\author{
Daniele Silvestro
}
