% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpAge.R
\name{simSpAge}
\alias{simSpAge}
\title{Simulation of Species Age}
\usage{
simSpAge(branching_time, lambda, mu, rho = 1, reps, batch = 100)
}
\arguments{
\item{branching_time}{time over which to run the simulation}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{rho}{sampling fraction (Default: 1.0)}

\item{reps}{number of stochastic ages}

\item{batch}{batch size of how many phylogenies are simualted at once}
}
\value{
A vector of stochastic species ages
}
\description{
This function stochastically simulates the distribution of species age
given its branching time, speciation and extinction rate, and the sampling fraction
}
\details{
Some additional details
}
\examples{
simSpAge(branching_time = 3.0, lambda = 0.2, mu = 0.1, rho = 0.7, reps = 10000)
}
\references{
Nee S., May R.M., and Harvey P.H. 1994.
The reconstructed evolutionary process.
Philos. Trans. R. Soc. Lond. B Biol. Sci. 344:305-311.
}
\seealso{
\code{\link{ageDensity}} for probabilistic age estimation
}
\author{
Daniele Silvestro
}
